Remove barcode.rs and create parse.rs
