+++ barkit-extract/src/parse.rs
@@ -1,8 +1,8 @@
-use crate::pattern;
-use regex::bytes::{Captures, Regex};
+use crate::pattern::BarcodeRegex;
+use regex::bytes::Captures;
 
 use seq_io::fastq::{OwnedRecord, Record, RefRecord};
-use std::{fmt, str};
+use std::str;
 
 use crate::error::Error;
 
@@ -28,76 +28,6 @@ const TRANSLATION_TABLE: [u8; 256] = {
 
     table
 };
-
-#[derive(Clone)]
-pub enum BarcodeType {
-    Umi,
-    Sample,
-    Cell,
-}
-
-impl fmt::Display for BarcodeType {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        let barcode_str = match self {
-            BarcodeType::Umi => "UMI",
-            BarcodeType::Sample => "SB",
-            BarcodeType::Cell => "CB",
-        };
-        write!(f, "{}", barcode_str)
-    }
-}
-
-impl BarcodeType {
-    fn get_barcode_type(name: &str) -> Result<Self, Error> {
-        match name {
-            "UMI" => Ok(BarcodeType::Umi),
-            "SB" => Ok(BarcodeType::Sample),
-            "CB" => Ok(BarcodeType::Cell),
-            _ => Err(Error::UnexpectedCaptureGroupName(name.to_owned())),
-        }
-    }
-}
-
-#[derive(Clone)]
-pub struct BarcodeRegex {
-    regex: Regex,
-    barcode_types: Vec<BarcodeType>,
-}
-
-impl BarcodeRegex {
-    pub fn new(pattern: &str, max_error: usize) -> Result<Self, Error> {
-        let fuzzy_pattern = pattern::get_with_errors(pattern, &max_error);
-        let regex = Regex::new(&fuzzy_pattern)?;
-        let barcode_types = Self::parse_capture_groups(&regex)?;
-        Ok(Self {
-            regex,
-            barcode_types,
-        })
-    }
-
-    pub fn get_captures<'a>(&'a self, read_seq: &'a [u8]) -> Result<Captures, Error> {
-        match self.regex.captures(read_seq) {
-            Some(capture) => Ok(capture),
-            None => Err(Error::PatternNotMatched),
-        }
-    }
-
-    fn parse_capture_groups(regex: &Regex) -> Result<Vec<BarcodeType>, Error> {
-        let mut capture_groups = Vec::<BarcodeType>::new();
-        for capture_group in regex
-            .capture_names()
-            .collect::<Vec<_>>()
-            .into_iter()
-            .flatten()
-        {
-            capture_groups.push(BarcodeType::get_barcode_type(capture_group)?)
-        }
-        if capture_groups.is_empty() {
-            return Err(Error::BarcodeCaptureGroupNotFound(regex.to_string()));
-        }
-        Ok(capture_groups)
-    }
-}
 
 pub struct BarcodeParser {
     barcode_regex: BarcodeRegex,
@@ -161,7 +91,7 @@ impl BarcodeParser {
         let seq = record.seq().to_vec();
         let qual = record.qual().to_vec();
 
-        for barcode in &self.barcode_regex.barcode_types {
+        for barcode in &self.barcode_regex.get_barcode_types() {
             let barcode_name = barcode.to_string();
             let (barcode_start, barcode_end) =
                 get_barcode_match_positions(&barcode_name, captures)?;
@@ -245,7 +175,7 @@ pub fn get_reverse_complement(sequence: &[u8]) -> Vec<u8> {
 mod tests {
     use rstest::rstest;
 
-    use crate::barcode::get_reverse_complement;
+    use crate::parse::get_reverse_complement;
 
     #[rstest]
     #[case(b"", b"")]
